Fix stack
Fixes the stack implementation, as it popped too early, a problem only
noticable with a loop depth greater than one (as `%TOP%` is preserved if
the stack is empty).
It also used to skip the first instruction after a loop (the character
after the `]`) if the loop (from the `[` character, skipping a loop
at `]` works fine) was skipped.